Ensure `Bump::new_chunk` can satisfy next request
This commit fixes a bug that was discovered through `wasm-bindgen` in
2.2.0 (maybe introduced in #5? unsure) where a call to `Bump::new_chunk`
may not have been guaranteed to satisfy the next requested allocation,
even though it was expected to be so.

The issue was that if a large layout was allocated when space was
reserved for `ChunkFooter` the allocation request didn't take that into
account. This means that if you requested a 10kb allocation, for
example, then after allocating a 10kb chunk the chunk wouldn't have 10kb
available because the footer was used as part of that 10kb.

The fix implemented in this issue is to ensure that the size is always
increased by the size of `ChunkFooter` if necessary, guaranteeing that
the chunk returned by `Bump::new_chunk` can satisfy the next allocation
if a layout is provided.