Fix a debug assert from firing in `new_chunk`
A debug assert was ensuring that the previous size specified for a chunk
was aligned to the chunk footer, but the previous change in #7 didn't
actually preserve that invariant when the user-provided size was used to
allocate a new chunk. While this doesn't actually cause an issues at
runtime (I don't think) it's probably best to go ahead and fix the issue
anyway by ensuring that the user-provided size, if used, is always aligned.