Make sure to align pointers in `alloc_layout_slow`
We didn't used to need to align pointers in `alloc_layout_slow` when we were
bumping upwards, since we made sure that the chunk itself had the proper
alignment, and we were using the start of the chunk for this allocation. Now
that we are bumping downwards, we need to round down to the proper alignment.

Fixes #43