Do not enable the `collections` cargo feature by default
It is incredibly hard to turn off default features. Every dependency in a crate
graph needs to disable them. Most people forget to disable default features they
aren't using and often aren't even aware that the features exist. So instead, it
is better to make people explicitly turn features on.