Add support for allocation limits on arenas
Arenas can optionally keep track of an `allocation_limit` which is
considered when allocating new chunks to an arena.

No new `new` methods were added, and limits must be set after the
creation of a `Bump` arena. Limits can be changed on the fly and are
only enforced when attempting to allocate new chunks.

No changes were made to the errors returned by operations like `alloc`
because of backwards compatibility.