Fix allocations on bumps with small limits exceeding limits
To ensure that allocations do not exceed the allocation limit imposed on
a `Bump`, the logic for calculating the final size of an allocated chunk
was factored out into a separate function so that the limit enforcing
code higher up in the control flow could more strictly enforce the
allocation limit, as previously this code did not have visibility into
the final size of an allocation request.

To ensure that allocations in arenas with small limits could succeed,
`alloc_layout_slow` was changed to allow bypassing the minimum chunk
size for new chunks when trying to allocate in new arenas when the limit
was lower than the default minimum chunk size.